+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), following
 
 ## [Unreleased]
 
+## [2.3.0] - 2026-08-18
+
 ### Removed
 
 - `conftest` binary from the `-circleci` image. It was only there for the `helm-conftest` step in
@@ -392,7 +394,8 @@ Initial release
     - config file is loaded from `.abs/main.yaml`, not from `.abs.yaml` (for future needs)
 - testing basic classes and pipelines
 
-[Unreleased]: https://github.com/giantswarm/app-build-suite/compare/v2.2.0...HEAD
+[Unreleased]: https://github.com/giantswarm/app-build-suite/compare/v2.3.0...HEAD
+[2.3.0]: https://github.com/giantswarm/app-build-suite/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/giantswarm/app-build-suite/compare/v2.1.3...v2.2.0
 [2.1.0]: https://github.com/giantswarm/app-build-suite/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/giantswarm/app-build-suite/compare/v1.8.1...v2.0.0